refactored articlesReducer
