+++ src/reducers/index.js
@@ -5,8 +5,7 @@ import { combineReducers } from 'redux'
 import userAuthReducer from './userAuthReducer'
 import addBoardReducer from './addBoardReducer'
 import getBoardsReducer from './getBoardsReducer'
-import addArticleReducer from './addArticleReducer'
-import getArticlesReducer from './getArticlesReducer'
+import articlesReducer from './articlesReducer'
 import deleteArticleReducer from './deleteArticleReducer'
 
 //root reducer
@@ -14,8 +13,7 @@ const rootReducer = combineReducers({
     userAuthReducer,
     addBoardReducer,
     getBoardsReducer,
-    addArticleReducer,
-    getArticlesReducer,
+    articlesReducer,
     deleteArticleReducer,
 })
